Move `useURLParam` to `hooks/`
