+++ src/hooks/index.tsx
@@ -1,3 +1,4 @@
 import { useAuth } from "./UseAuth";
+import { useURLParam } from "./UseURLParam";
 
-export { useAuth };
+export { useAuth, useURLParam };
