+++ src/pages/Statuses.tsx
@@ -1,37 +1,10 @@
 import { useNavigate } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
 import { Box, Container, Typography } from "@mui/material";
+
+import { useURLParam } from "../hooks";
 import { StatusEnum, useStatusesQuery } from "../graph";
-
-interface ConvertToString<U> {
-  (value: U): string;
-}
-
-interface ConvertFromString<U> {
-  (value: string): U;
-}
-
-function useURLParam<U>(
-  urlSearchParams: URLSearchParams,
-  key: string,
-  convertToString: ConvertToString<U>,
-  convertFromString: ConvertFromString<U>
-): [U | undefined, React.Dispatch<React.SetStateAction<U | undefined>>] {
-  const raw = urlSearchParams.get(key);
-  const [param, setParam] = useState<U | undefined>(
-    raw === null ? undefined : convertFromString(raw)
-  );
-  useEffect(() => {
-    if (param === undefined) {
-      urlSearchParams.delete(key);
-    } else {
-      urlSearchParams.set(key, convertToString(param));
-    }
-  }, [param]);
-
-  return [param, setParam];
-}
 
 export default function Statuses() {
   const navigate = useNavigate();
